+++ node/params.d.ts
@@ -1,7 +1,0 @@
-type Params = {
-    paramsList: string;
-    paramsNameAppend: string;
-    splitparams: string;
-};
-declare const $urlParams: ({ paramsList, splitparams, paramsNameAppend }: Params) => URLSearchParams;
-export { $urlParams };